// src/app/modules/logistica/entrada-materiais/parecer/models/parecer.ts
export interface ILogisticaEntradaMateriaisParecer {
  ID_LOGI_ENMA_FHNC_OCPR:   number;
  NM_PARE:                  number;
  IN_STAT:                  number; 
  DS_OBSE:                  string;
  NR_MATR:                  number;
  NM_USUA:                  string;
  DT_INCL:                  string; 
  DT_ATUA:                  string;
  ID_USUA:                  number;
}